requirements.txt
# Core dependencies
fastapi==0.135.3
uvicorn==0.42.0
python-dotenv==1.2.2

# Database
SQLAlchemy==2.0.49

# Data validation
pydantic==2.12.5
email-validator==2.3.0

# Security & Authentication
PyJWT==2.12.1
pwdlib==0.3.0

# HTTP parsing
httptools==0.7.1
watchfiles==1.1.1
